Perlin noise to use floats instead of doubles

+++ perlin_noise.cc
@@ -42,7 +42,7 @@ PerlinNoise::PerlinNoise(unsigned int seed) {
 	p.insert(p.end(), p.begin(), p.end());
 }
 
-double PerlinNoise::noise(double x, double y, double z) const {
+float PerlinNoise::noise(float x, float y, float z) const {
 	// Find the unit cube that contains the point
 	int X = (int) floor(x) & 255;
 	int Y = (int) floor(y) & 255;
@@ -54,9 +54,9 @@ double PerlinNoise::noise(double x, double y, double z) const {
 	z -= floor(z);
 
 	// Compute fade curves for each of x, y, z
-	double u = fade(x);
-	double v = fade(y);
-	double w = fade(z);
+	float u = fade(x);
+	float v = fade(y);
+	float w = fade(z);
 
 	// Hash coordinates of the 8 cube corners
 	int A = p[X] + Y;
@@ -67,22 +67,22 @@ double PerlinNoise::noise(double x, double y, double z) const {
 	int BB = p[B + 1] + Z;
 
 	// Add blended results from 8 corners of cube
-	double res = lerp(w, lerp(v, lerp(u, grad(p[AA], x, y, z), grad(p[BA], x-1, y, z)), lerp(u, grad(p[AB], x, y-1, z), grad(p[BB], x-1, y-1, z))),	lerp(v, lerp(u, grad(p[AA+1], x, y, z-1), grad(p[BA+1], x-1, y, z-1)), lerp(u, grad(p[AB+1], x, y-1, z-1),	grad(p[BB+1], x-1, y-1, z-1))));
+	float res = lerp(w, lerp(v, lerp(u, grad(p[AA], x, y, z), grad(p[BA], x-1, y, z)), lerp(u, grad(p[AB], x, y-1, z), grad(p[BB], x-1, y-1, z))),	lerp(v, lerp(u, grad(p[AA+1], x, y, z-1), grad(p[BA+1], x-1, y, z-1)), lerp(u, grad(p[AB+1], x, y-1, z-1),	grad(p[BB+1], x-1, y-1, z-1))));
 	return (res + 1.0)/2.0;
 }
 
-double PerlinNoise::fade(double t) const {
+float PerlinNoise::fade(float t) const {
 	return t * t * t * (t * (t * 6 - 15) + 10);
 }
 
-double PerlinNoise::lerp(double t, double a, double b) const {
+float PerlinNoise::lerp(float t, float a, float b) const {
 	return a + t * (b - a);
 }
 
-double PerlinNoise::grad(int hash, double x, double y, double z) const {
+float PerlinNoise::grad(int hash, float x, float y, float z) const {
 	int h = hash & 15;
 	// Convert lower 4 bits of hash into 12 gradient directions
-	double u = h < 8 ? x : y,
+	float u = h < 8 ? x : y,
 		   v = h < 4 ? y : h == 12 || h == 14 ? x : z;
 	return ((h & 1) == 0 ? u : -u) + ((h & 2) == 0 ? v : -v);
 }
