+++ perlin_noise.h
@@ -12,11 +12,11 @@ public:
 	// Generate a new permutation vector based on the value of seed
 	PerlinNoise(unsigned int seed);
 	// Get a noise value, for 2D images z can have any value
-	double noise(double x, double y, double z) const;
+	float noise(float x, float y, float z) const;
 private:
-	double fade(double t) const;
-	double lerp(double t, double a, double b) const;
-	double grad(int hash, double x, double y, double z) const;
+	float fade(float t) const;
+	float lerp(float t, float a, float b) const;
+	float grad(int hash, float x, float y, float z) const;
 };
 
 #endif
